Refactored project packaging
