+++ Core/Structures/BloomFilter/BFCalculations.go
