+++ Core/Structures/SkipList/test.go
